Add utils as an import to tests and remove whitespacing
