+++ tests/firedrake/regression/test_auxiliary_dm.py
@@ -1,5 +1,6 @@
 from firedrake import *
 from firedrake.petsc import DEFAULT_DIRECT_SOLVER_PARAMETERS
+from firedrake.utils import complex_mode
 import pytest
 distribution_parameters = {"partition": True, "overlap_type": (DistributedMeshOverlapType.VERTEX, 2)}
 
@@ -45,7 +46,7 @@ class BiharmonicProblem(object):
         return f
 
 
-@pytest.mark.skipif(utils.complex_mode, reason="Differentiation of energy not defined in Complex.")
+@pytest.mark.skipif(complex_mode, reason="Differentiation of energy not defined in Complex.")
 def test_auxiliary_dm():
     problem = BiharmonicProblem(5, 1)
     mesh = problem.mesh()
