+++ tests/firedrake/regression/test_coefficient_derivatives.py
@@ -1,9 +1,10 @@
 from firedrake import *
+from firedrake.utils import complex_mode
 import numpy as np
 import pytest
 
 
-@pytest.mark.skipif(utils.complex_mode, reason="Not complex differentiable")
+@pytest.mark.skipif(complex_mode, reason="Not complex differentiable")
 def test_coefficient_derivatives():
     m = UnitSquareMesh(3, 3)
     x = SpatialCoordinate(m)
