+++ tests/firedrake/regression/test_coordinatederivative.py
@@ -1,9 +1,10 @@
 import pytest
 import numpy as np
 from firedrake import *
+from firedrake.utils import complex_mode
 
 
-@pytest.mark.skipif(utils.complex_mode, reason="Don't expect coordinate derivatives to work in complex")
+@pytest.mark.skipif(complex_mode, reason="Don't expect coordinate derivatives to work in complex")
 def test_first_shape_derivative():
     mesh = UnitSquareMesh(6, 6)
     n = FacetNormal(mesh)
@@ -47,7 +48,7 @@ def test_first_shape_derivative():
     test_first(J, dJ)
 
 
-@pytest.mark.skipif(utils.complex_mode, reason="Don't expect coordinate derivatives to work in complex")
+@pytest.mark.skipif(complex_mode, reason="Don't expect coordinate derivatives to work in complex")
 def test_mixed_derivatives():
     mesh = UnitSquareMesh(6, 6)
     X = SpatialCoordinate(mesh)
@@ -80,7 +81,7 @@ def test_mixed_derivatives():
     test_mixed(J, dJ)
 
 
-@pytest.mark.skipif(utils.complex_mode, reason="Don't expect coordinate derivatives to work in complex")
+@pytest.mark.skipif(complex_mode, reason="Don't expect coordinate derivatives to work in complex")
 def test_integral_scaling_edge_case():
     mesh = UnitSquareMesh(6, 6)
     X = SpatialCoordinate(mesh)
@@ -96,7 +97,7 @@ def test_integral_scaling_edge_case():
         assemble(Constant(2.0) * derivative(derivative(J, X), X))
 
 
-@pytest.mark.skipif(utils.complex_mode, reason="Don't expect coordinate derivatives to work in complex")
+@pytest.mark.skipif(complex_mode, reason="Don't expect coordinate derivatives to work in complex")
 def test_second_shape_derivative():
     mesh = UnitSquareMesh(6, 6)
     V = FunctionSpace(mesh, "CG", 1)
@@ -129,7 +130,7 @@ def test_second_shape_derivative():
     test_second(Ja+Jb, ddJa + ddJb)
 
 
-@pytest.mark.skipif(utils.complex_mode, reason="Don't expect coordinate derivatives to work in complex")
+@pytest.mark.skipif(complex_mode, reason="Don't expect coordinate derivatives to work in complex")
 def test_coordinate_handling():
     mesh = UnitSquareMesh(60, 60)
     V = VectorFunctionSpace(mesh, "CG", 1)
